backend/id_checker_project/requirements.txt
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
gunicorn==20.1.0
mysqlclient==2.2.5              
requests==2.31.0
